+++ test_desktop/data/mbus_decoder_test.cpp
@@ -18,7 +18,7 @@ namespace mbus_decoder_test {
                 return 0;
             }
         
-            size_t getFrameSize() const override {
+            int getFrameSize() const override {
                 return size_;
             }
